pyproject.toml
[tool.poetry]
name = "nothing-filter"
version = "0.1.0"
description = ""
authors = ["maxomm <max.warsinke@googlemail.com>"]

[tool.poetry.dependencies]
python = "3.10.8"
pillow = "^9.3.0"
streamlit = "^1.14.0"
numpy = "^1.23.4"


[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
flake8 = "^5.0.4"
pylint = "^2.15.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
